Organize project (#3)
* Organizing project and displaying initial components

Moved a lot of files to have an initial project structure. The project's now rendering a search bar.

* Display card (without styles)

A card is being displayed, no CSS. Some more files were added to finally use the components.

* Retrieving data and showing it

The API is making calls now. The Home component is rendering the default search.

* Display the first part of the VideoDetail page

* Finally showing video's information on VideoDetail

After many attempts, the information for the selected video is now showing in the desired component.

* Removing old paths

* Re-adding paths and proptypes

* Rename file

* Redirection needed for Netlify