VideoDetail info not showing up
I thought this component was working fine, but when clicking on related videos I noticed that the logic was not correct at all. This commit tries the approach of storing the selectedVideo on the Context API to later consume it on another component, but doesn't seem to work at all.